Remove the dependency on the epel-release package (PR #161)
# Description

Remove the dependency on the epel-release package. This package is required to be able to install python3.13. But adding it as a dependency of the inmanta RPMs is not a good solution, because that package just adds a yum `.repo` file. Hence epel-release has to be installed before executing the `dnf install inmanta-service-orchestrator-server` command.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry